test: move all nonlocal tests to test files excluded by --tags=local_tests
