+++ server/keyshare/tasks/tasks_test.go
@@ -128,6 +128,70 @@ func TestExpireAccounts(t *testing.T) {
 	assert.Equal(t, 1, countRows(t, db, "users", "delete_on IS NOT NULL"))
 }
 
+func TestConfiguration(t *testing.T) {
+	testdataPath := test.FindTestdataFolder(t)
+
+	err := processConfiguration(&Configuration{Logger: irma.Logger})
+	assert.NoError(t, err)
+
+	err = processConfiguration(&Configuration{
+		EmailConfiguration: keyshare.EmailConfiguration{
+			EmailServer:     "localhost:1025",
+			EmailFrom:       "test@test.com",
+			DefaultLanguage: "en",
+		},
+		DeleteExpiredAccountFiles: map[string]string{
+			"en": filepath.Join(testdataPath, "emailtemplate.html"),
+		},
+		DeleteExpiredAccountSubjects: map[string]string{
+			"en": "testsubject",
+		},
+		Logger: irma.Logger,
+	})
+	assert.NoError(t, err)
+
+	err = processConfiguration(&Configuration{
+		EmailConfiguration: keyshare.EmailConfiguration{
+			EmailServer: "localhost:1025",
+			EmailFrom:   "test@test.com",
+		},
+		DeleteExpiredAccountFiles: map[string]string{
+			"en": filepath.Join(testdataPath, "emailtemplate.html"),
+		},
+		DeleteExpiredAccountSubjects: map[string]string{
+			"en": "testsubject",
+		},
+		Logger: irma.Logger,
+	})
+	assert.Error(t, err)
+
+	err = processConfiguration(&Configuration{
+		EmailConfiguration: keyshare.EmailConfiguration{
+			EmailServer:     "localhost:1025",
+			EmailFrom:       "test@test.com",
+			DefaultLanguage: "en",
+		},
+		DeleteExpiredAccountSubjects: map[string]string{
+			"en": "testsubject",
+		},
+		Logger: irma.Logger,
+	})
+	assert.Error(t, err)
+
+	err = processConfiguration(&Configuration{
+		EmailConfiguration: keyshare.EmailConfiguration{
+			EmailServer:     "localhost:1025",
+			EmailFrom:       "test@test.com",
+			DefaultLanguage: "en",
+		},
+		DeleteExpiredAccountFiles: map[string]string{
+			"en": filepath.Join(testdataPath, "emailtemplate.html"),
+		},
+		Logger: irma.Logger,
+	})
+	assert.Error(t, err)
+}
+
 func SetupDatabase(t *testing.T) {
 	test.RunScriptOnDB(t, "../cleanup.sql", true)
 	test.RunScriptOnDB(t, "../schema.sql", false)
